refactor(Structure): Move redux related folders into own dir
