+++ src/client/redux/reducers/ThreadMonitor.js
@@ -1,9 +1,9 @@
-import initialState from '~/constants/initialState';
+import initialState from '../initialState';
 import { 
     THREAD_MONITOR_ADDED,
     THREAD_MONITOR_DELETED,
     THREAD_MONITOR_UPDATED,
-} from '~/constants'
+} from '../types'
 
 // TODO: when a user inspects new posts - a counter of some sort; notifications
